Make fields const and more data to proxy Mat

+++ src/skepu3/cluster/gpi/matrix.hpp
@@ -71,7 +71,7 @@ namespace skepu{
     T* local_buffer;
 
     int local_size;
-    long global_size;
+    size_t global_size;
 
     unsigned long comm_offset;
     unsigned long comm_size;
